Refaktoryzacja konfiguracji sprzętu Raspberry Pi
Podzielono ConfigureRaspberryPi na osobne klasy: ConfigureSystemd, ConfigureAvahi, ConfigureI2C, każda implementuje nowy interfejs ISystemConfigurer. Dodano RaspberryPiCommandHelper do obsługi poleceń systemowych. HardwareCheckInterceptor obsługuje teraz wiele konfigurerów. Zaktualizowano rejestrację DI oraz rozszerzono interfejs o Name i CanConfigure.
