+++ src/ZtrBoardGame.Console/Infrastructure/HardwareCheckInterceptor.cs
@@ -2,14 +2,15 @@ using Microsoft.Extensions.Options;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using System;
+using System.Collections.Generic;
 using System.Threading;
 using ZtrBoardGame.Configuration.Shared;
 using ZtrBoardGame.Console.Commands.Board;
-using ZtrBoardGame.RaspberryPi.HardwareAccess;
+using ZtrBoardGame.RaspberryPi.HardwareAccess.SystemConfigurers;
 
 namespace ZtrBoardGame.Console.Infrastructure;
 
-class HardwareCheckInterceptor(IAnsiConsole console, IOptions<HardwareConfigurationSettings> config, ISystemConfigurer systemConfigurer) : ICommandInterceptor
+class HardwareCheckInterceptor(IAnsiConsole console, IOptions<HardwareConfigurationSettings> config, IEnumerable<ISystemConfigurer> systemConfigurers) : ICommandInterceptor
 {
     public void Intercept(CommandContext context, CommandSettings settings)
     {
@@ -18,50 +19,54 @@ class HardwareCheckInterceptor(IAnsiConsole console, IOptions<HardwareConfigurat
             return;
         }
 
-        var isConfigNeeded = false;
-        try
+        foreach (var systemConfigurer in systemConfigurers)
         {
-            isConfigNeeded = systemConfigurer.IsConfigurationNeeded();
-        }
-        catch (Exception e)
-        {
-            console.WriteException(e);
-        }
 
-        if (!isConfigNeeded)
-        {
-            return;
-        }
-
-        console.Write(new Rule("[yellow]Hardware configuration is required.[/]"));
-
-        var confirm = true;
-        if (!config.Value.DoAutoConfig)
-        {
-            confirm = console.Confirm("Do you want to configure the system now?");
-        }
-
-        if (confirm)
-        {
+            var isConfigNeeded = false;
             try
             {
-                AnsiConsole.Status()
-                    .Start("Configuring RaspberryPi...", ctx =>
-                    {
-                        systemConfigurer.Configure();
-                        Thread.Sleep(500);
-                    });
-
-                console.MarkupLine("[green]System configured successfully.[/]");
+                isConfigNeeded = systemConfigurer.IsConfigurationNeeded();
             }
             catch (Exception e)
             {
                 console.WriteException(e);
             }
-        }
-        else
-        {
-            console.MarkupLine("[red]System configuration skipped. The application may not function correctly.[/]");
+
+            if (!isConfigNeeded)
+            {
+                return;
+            }
+
+            console.Write(new Rule($"[yellow]Hardware configuration for {systemConfigurer.Name} is required.[/]"));
+
+            var confirm = true;
+            if (!config.Value.DoAutoConfig)
+            {
+                confirm = console.Confirm($"Do you want to configure the {systemConfigurer.Name} now?");
+            }
+
+            if (confirm)
+            {
+                try
+                {
+                    AnsiConsole.Status()
+                        .Start("Configuring RaspberryPi...", ctx =>
+                        {
+                            systemConfigurer.Configure();
+                            Thread.Sleep(500);
+                        });
+
+                    console.MarkupLine($"[green]{systemConfigurer.Name} configured successfully.[/]");
+                }
+                catch (Exception e)
+                {
+                    console.WriteException(e);
+                }
+            }
+            else
+            {
+                console.MarkupLine($"[red]{systemConfigurer.Name} configuration skipped. The application may not function correctly.[/]");
+            }
         }
     }
 }
