+++ src/ZtrBoardGame.RaspberryPi/HardwareAccess/SystemConfigurers/ConfigureI2C.cs
@@ -1,0 +1,32 @@
+﻿namespace ZtrBoardGame.RaspberryPi.HardwareAccess.SystemConfigurers;
+
+class ConfigureI2C() : ISystemConfigurer
+{
+    public string Name => "I2C Bus";
+
+    public bool CanConfigure()
+        => RaspberryPiCommandHelper.IsRaspberryPi();
+
+    public bool IsConfigurationNeeded()
+        => !IsI2CEnabled();
+
+    public void Configure()
+        => EnableI2C();
+
+    private static bool IsI2CEnabled()
+    {
+        try
+        {
+            var output = RaspberryPiCommandHelper.RunCommand("raspi-config", "nonint get_i2c", "Cannot check if I2C bus is enabled",
+                redirectStandardOutput: true);
+            return output == "0";
+        }
+        catch
+        {
+            return false;
+        }
+    }
+
+    private static void EnableI2C()
+        => RaspberryPiCommandHelper.RunCommand("raspi-config", "nonint do_i2c 0", "Cannot enable I2C bus", redirectStandardOutput: true);
+}
