+++ src/ZtrBoardGame.RaspberryPi/HardwareAccess/SystemConfigurers/ISystemConfigurer.cs
@@ -1,0 +1,9 @@
+﻿namespace ZtrBoardGame.RaspberryPi.HardwareAccess.SystemConfigurers;
+
+public interface ISystemConfigurer
+{
+    bool IsConfigurationNeeded();
+    bool CanConfigure();
+    void Configure();
+    string Name { get; }
+}
